+++ Assets/Demos/Pong/Core.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 86d6a5a6356f1406881eaad66fb770f9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
