+++ Assets/Demos/Pong/Core/Globals.cs
