+++ Assets/Demos/Pong/Network/Client.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f7d79b8c9b80b4828a30c9d1e63879b8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
