+++ Assets/Demos/Pong/Network/Client/PaddleLeftSyncClient.cs
