+++ Assets/Demos/Pong/Network/Client/PaddleRightSyncClient.cs
