+++ Assets/Demos/Pong/Network/Server/BallSyncServer.cs
@@ -1,11 +1,5 @@
 using UnityEngine;
 
-[System.Serializable]
-public class BallState {
-    public Vector3 Position;
-}
-
-
 public class BallSyncServer : MonoBehaviour
 {
     ServerManager ServerMan;
