+++ Assets/Demos/Pong/States/BallState.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+[System.Serializable]
+public class BallState {
+    public Vector3 Position;
+}
