+++ Assets/Demos/Pong/States/PaddleLeftState.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+[System.Serializable]
+public class PaddleLeftState {
+    public Vector3 Position;
+}
