+++ Assets/Demos/Pong/States/PaddleRightState.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+[System.Serializable]
+public class PaddleRightState {
+    public Vector3 Position;
+}
