+++ Assets/Demos/Pong/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 25023bda7f0924ff9955f21532e991c5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
